Update version, begin TypeScript conversion, Begin option screen refactor
