+++ src/components/shared/StickySpacer.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const StickySpacer = () => {
+    return <div
+    style={{
+        position: "sticky",
+        bottom: "0px",
+        height: "25vh",
+        pointerEvents: "none",
+    }}
+/>
+}
+
+export default StickySpacer;
